+++ src/app/public/modules/lookup/lookup.component.spec.ts
@@ -71,6 +71,12 @@ describe('Lookup component', function () {
     tick();
   }
 
+  function clickSearchButton(fixture: ComponentFixture<any>): void {
+    getSearchButton().click();
+    fixture.detectChanges();
+    tick();
+  }
+
   function clickShowMoreClearAll(fixture: ComponentFixture<any>): void {
     (<HTMLElement>document.querySelector('.sky-lookup-show-more-modal-clear-all-btn')).click();
     fixture.detectChanges();
@@ -103,23 +109,37 @@ describe('Lookup component', function () {
   }
 
   function getAddButton(): HTMLElement {
-    return document.querySelector('.sky-autocomplete-add') as HTMLElement;
+    return document.querySelector('.sky-autocomplete-action-add') as HTMLElement;
+  }
+
+  function getDropdown(): HTMLElement {
+    return document.querySelector('.sky-autocomplete-results-container');
   }
 
   function getInputElement(lookupComponent: SkyLookupComponent): HTMLInputElement {
-    return lookupComponent['elementRef'].nativeElement.querySelector('.sky-lookup-input');
+    return lookupComponent['lookupWrapperRef'].nativeElement.querySelector('.sky-lookup-input');
   }
 
   function getModalAddButton(): HTMLElement {
     return document.querySelector('.sky-lookup-show-more-modal-add') as HTMLElement;
   }
 
+  function getModalSearchInputValue(): string {
+    const modalSearchInput =
+      document.querySelector('.sky-search-input-container .sky-form-control') as HTMLInputElement;
+    return modalSearchInput.value;
+  }
+
   function getRepeaterItemCount(): number {
     return document.querySelectorAll('sky-modal sky-repeater-item').length;
   }
 
+  function getSearchButton(): HTMLElement {
+    return document.querySelector('.sky-input-group-btn .sky-btn') as HTMLElement;
+  }
+
   function getShowMoreButton(): HTMLElement {
-    return document.querySelector('.sky-autocomplete-more') as HTMLElement;
+    return document.querySelector('.sky-autocomplete-action-more') as HTMLElement;
   }
 
   function getShowMoreRepeaterItemContent(index: number): string {
@@ -377,7 +397,7 @@ describe('Lookup component', function () {
         }));
 
         it('should not do anything on token click', fakeAsync(() => {
-          const showMoreSpy = spyOn(component.lookupComponent, 'showMoreButtonClicked').and.stub();
+          const showMoreSpy = spyOn(component.lookupComponent, 'openPicker').and.stub();
 
           component.friends = [
             { name: 'Fred' },
@@ -645,6 +665,7 @@ describe('Lookup component', function () {
 
             addButton.click();
             fixture.detectChanges();
+            tick();
 
             expect(addButtonSpy).toHaveBeenCalled();
           })
@@ -699,6 +720,52 @@ describe('Lookup component', function () {
             clickShowMore(fixture);
 
             expect(modalService.open).toHaveBeenCalled();
+
+            closeModal(fixture);
+          })
+        );
+
+        it('should open the modal when the search button is clicked',
+          fakeAsync(() => {
+            component.enableShowMore = true;
+            fixture.detectChanges();
+
+            spyOn(modalService, 'open').and.callThrough();
+
+            clickSearchButton(fixture);
+
+            expect(modalService.open).toHaveBeenCalled();
+
+            closeModal(fixture);
+          })
+        );
+
+        it('should close the dropdown when the search button is clicked',
+          fakeAsync(() => {
+            component.enableShowMore = true;
+            fixture.detectChanges();
+
+            performSearch('r', fixture);
+            expect(getDropdown()).not.toBeNull();
+
+            clickSearchButton(fixture);
+
+            expect(getDropdown()).toBeNull();
+
+            closeModal(fixture);
+          })
+        );
+
+        it('should populate search bar with current input value when the search button is clicked',
+          fakeAsync(() => {
+            component.enableShowMore = true;
+            fixture.detectChanges();
+
+            performSearch('foo', fixture);
+
+            clickSearchButton(fixture);
+
+            expect(getModalSearchInputValue()).toEqual('foo');
 
             closeModal(fixture);
           })
@@ -998,7 +1065,7 @@ describe('Lookup component', function () {
           }));
 
           it('should open the show more modal on token click', fakeAsync(() => {
-            const showMoreSpy = spyOn(component.lookupComponent, 'showMoreButtonClicked').and.stub();
+            const showMoreSpy = spyOn(component.lookupComponent, 'openPicker').and.stub();
 
             component.enableShowMore = true;
             component.friends = [
@@ -1013,7 +1080,7 @@ describe('Lookup component', function () {
           }));
 
           it('should open the show more modal on collapsed token click', fakeAsync(() => {
-            const showMoreSpy = spyOn(component.lookupComponent, 'showMoreButtonClicked').and.stub();
+            const showMoreSpy = spyOn(component.lookupComponent, 'openPicker').and.stub();
 
             component.enableShowMore = true;
             component.friends = [
@@ -1818,7 +1885,7 @@ describe('Lookup component', function () {
         }));
 
         it('should not do anything on token click', fakeAsync(() => {
-          const showMoreSpy = spyOn(component.lookupComponent, 'showMoreButtonClicked').and.stub();
+          const showMoreSpy = spyOn(component.lookupComponent, 'openPicker').and.stub();
 
           component.selectedFriends = [
             { name: 'Fred' },
@@ -1950,6 +2017,7 @@ describe('Lookup component', function () {
 
             addButton.click();
             fixture.detectChanges();
+            tick();
 
             expect(addButtonSpy).toHaveBeenCalled();
           })
@@ -2304,7 +2372,7 @@ describe('Lookup component', function () {
           }));
 
           it('should open the show more modal on token click', fakeAsync(() => {
-            const showMoreSpy = spyOn(component.lookupComponent, 'showMoreButtonClicked').and.stub();
+            const showMoreSpy = spyOn(component.lookupComponent, 'openPicker').and.stub();
 
             component.enableShowMore = true;
             component.selectedFriends = [
@@ -2321,7 +2389,7 @@ describe('Lookup component', function () {
           }));
 
           it('should open the show more modal on collapsed token click', fakeAsync(() => {
-            const showMoreSpy = spyOn(component.lookupComponent, 'showMoreButtonClicked').and.stub();
+            const showMoreSpy = spyOn(component.lookupComponent, 'openPicker').and.stub();
 
             component.enableShowMore = true;
             component.selectedFriends = [
@@ -2942,9 +3010,13 @@ describe('Lookup component', function () {
   describe('inside input box', () => {
     let fixture: ComponentFixture<SkyLookupInputBoxTestComponent>;
     let nativeElement: HTMLElement;
+    let component: SkyLookupInputBoxTestComponent;
+    let lookupComponent: SkyLookupComponent;
 
     beforeEach(() => {
       fixture = TestBed.createComponent(SkyLookupInputBoxTestComponent);
+      component = fixture.componentInstance;
+      lookupComponent = component.lookupComponent;
       nativeElement = fixture.nativeElement as HTMLElement;
     });
 
@@ -2959,5 +3031,50 @@ describe('Lookup component', function () {
       expect(containerEl).toHaveCssClass('sky-lookup');
     }));
 
+    it('should unfocus the component if it loses focus', fakeAsync(function () {
+      fixture.detectChanges();
+
+      const inputElement = getInputElement(lookupComponent);
+      SkyAppTestUtility.fireDomEvent(inputElement, 'focusin');
+      tick();
+      fixture.detectChanges();
+      tick();
+
+      expect(lookupComponent.isInputFocused).toEqual(true);
+
+      SkyAppTestUtility.fireDomEvent(document, 'focusin');
+      tick();
+      fixture.detectChanges();
+      tick();
+
+      expect(lookupComponent.isInputFocused).toEqual(false);
+    }));
+
+    describe('mouse interactions', function () {
+      it('should focus the input if the host is clicked', fakeAsync(function () {
+        fixture.detectChanges();
+
+        const hostElement = document.querySelector('.sky-lookup');
+        const input = getInputElement(lookupComponent);
+
+        triggerClick(hostElement, fixture);
+
+        expect(document.activeElement).toEqual(input);
+      }));
+
+      it('should not focus the input if a token is clicked', fakeAsync(function () {
+        fixture.detectChanges();
+
+        performSearch('s', fixture);
+        selectSearchResult(0, fixture);
+
+        const tokenElements = getTokenElements();
+        const input = getInputElement(lookupComponent);
+
+        triggerClick(tokenElements.item(0), fixture, true);
+
+        expect(document.activeElement).not.toEqual(input);
+      }));
+    });
   });
 });